+++ packages/sdk/src/factory/chain-provider.ts
@@ -1,4 +1,5 @@
 import { ethers } from "ethers";
+import { CHAIN_IDS } from "../common/chain-ids";
 import { Config } from "../common/config";
 
 
@@ -20,11 +21,18 @@ export function createChainProvider(chainId: number) {
       quorum: Config.QUORUM,
     }),
   };
-  
-  const chainProvider = ethers.getDefaultProvider(
-    chainId,
-    chainProviderOptions
-  );
+
+  let chainProvider;
+  if (chainId === CHAIN_IDS.BNB_MAINNET) {
+    chainProvider = new ethers.providers.JsonRpcProvider('https://bsc-dataseed.binance.org/', 56);
+  } else {
+    chainProvider = ethers.getDefaultProvider(
+      chainId,
+      chainProviderOptions
+    );
+  }
+
+
   chainProviderCache[chainId] = chainProvider;
 
   return chainProvider;